Make `request.options` & `request.registered_options` internals private
Making internal representation private gives us room to manoeuvre in the
future. For example, changing options to keywords list would allow
keeping duplicate keys and these could be handled on a per-step basis,
for example `put_params/1` step could _merge_ params when someone wrote
`Req.new(params: [a: 1]) |> Req.update(params: [b: 2])` as opposed to
currently taking the _latter_.

Another reason is, with maps developers might be tempted to write:

    def my_step(req) when req.options.foo == true

which is perfectly fine. However, changing this to:

    def my_step(req) when req.options.foo == true or req.method == :get

is a bug. This code would work if `req.options` contains `:foo` key
but if it doesn't, the ` or req.method == :get` is never executed.
Changing the order would fix the problem:

    def my_step(req) when req.method == :get or req.options.foo == true

but this may not be immediately obvious.