Add support for deflate decompression
This adds support for responses that were compressed using the `deflate` algorithm.

As best I can tell, this isn't terribly common for servers to send in the wild—in my own usage of Req for scraping public sites, I see gzip and Brotli way, way more frequently. However, I think it's useful to have in Req for the sake of having decompression support that matches real browsers; Safari, Firefox, and Chrome each send an `Accept-Encoding` header of `gzip, deflate, br`.