Automatically parse request body when using `:plug` option
To match automatically fetching the query_params, the request body is
now also parsed when using the `:plug` option in the request. This is
available on `conn.body_params`. The raw body is also stored as a
binary, which can be read using the new `Req.Test.raw_body/1` function.

This simplifies some of the tests which were calling `Plug.Parsers`
manually.