Use a custom adapter for transitioning to automatic body reads
In commit c17cd20 using `run_plug` was modified to ensure that the body
was read. This was much more convenient for tests that assert on
body_params for JSON or file uploads, however it could break existing
tests that were using `read_body` directly.

This commit adds a custom adapter that supports calling read_body,
however it warns the user. The body is cached in the custom adapter, and
re-populates the body when reading, to ensure that the existing
`read_body` options work.

In a future commit, we should add the ability to opt-out of the
automatic parsing, which would inform the custom adapter to simply
delegate to the `Plug.Adapters.Test.Conn` adapter.